+++ tests/malloc_test/central_free_list.h
@@ -11,31 +11,31 @@ public:
 	~CentralFreeList();
 
 public:
-	s32 Init(SizeClass cl);
-	s32 InsertRange(void* start, void* end, s32 num);
-	s32 RemoveRange(void** start, void** end, s32 num);
+	int Init(SizeClass cl);
+	int InsertRange(void* start, void* end, int num);
+	int RemoveRange(void** start, void** end, int num);
 
 	// 返回从span中取得了多少对象
-	s32 FetchFromOneSpan(s32 num, void** start, void** end);
+	int FetchFromOneSpan(int num, void** start, void** end);
 	// 先尝试从span获取。span中不足，从page heap分配span,然后再获取
-	s32 FetchFromOneSpanSafe(s32 num, void** start, void** end);
+	int FetchFromOneSpanSafe(int num, void** start, void** end);
 
-	s32 ReleaseListToSpans(void* start);
-	s32 ReleaseToSpans(void* object);
+	int ReleaseListToSpans(void* start);
+	int ReleaseToSpans(void* object);
 
 	// 从page heap 中分配span,切分为cache对象
-	s32 Populate();
+	int Populate();
 
 	// 试图增大tc entry cache的数量
 	bool TryMakeCacheSpace();
 	// 收缩cache大小
-	s32 ShrinkCache(bool is_force);
+	int ShrinkCache(bool is_force);
 
 	// 统计所有碎片大小，数量的总和
-	u64 OverheadBytes() const;
-	s32 TCTotalLength() const;
+	unsigned long long OverheadBytes() const;
+	int TCTotalLength() const;
 
-	u64 free_obj_num() const { return free_obj_num_; }
+	unsigned long long free_obj_num() const { return free_obj_num_; }
 
 	static bool EvictRandomSizeClass(SizeClass lock_cl, bool is_force);
 
@@ -47,16 +47,16 @@ public:
 	};
 
 	// 最大允许缓存多少条链表
-	static const s32 kMaxNumTransferEntries = 64;
+	static const int kMaxNumTransferEntries = 64;
 
 private:
 	SizeClass cl_;
 	Span empty_; // 已经全部分配出去的span
 	Span noempty_; // 部分分配出去的span, 还可以继续分配
-	u64 num_spans_; // empty + noempty的span总和
+	unsigned long long num_spans_; // empty + noempty的span总和
 	TCEntry tc_slots_[kMaxNumTransferEntries];
-	s32 used_slots_; // 使用的tc_slots数目
-	s32 cache_size_; // 当前允许缓存slot大小
-	s32 max_cache_size_; // 最大允许缓存slot大小
-	u64 free_obj_num_; // 空闲对象数量,包含span内的对象数量
+	int used_slots_; // 使用的tc_slots数目
+	int cache_size_; // 当前允许缓存slot大小
+	int max_cache_size_; // 最大允许缓存slot大小
+	unsigned long long free_obj_num_; // 空闲对象数量,包含span内的对象数量
 };
